Chart Sample
Starting up the app checks to see whether the database (chart.sqlite)
and the table name 'water' otherwise creates them.
Following at least two records are added to the database through the
action "Add" to generate and display the graph.
Holding down on a list item, record the Db, you activate the context
menu, and you can delete it.
 
Functionality: Used
- Databases: Creation and connection
- Table: creation, connection, adding and deleting records
- Graphics: create, update, and display
- Toast messages
- Dialogs: prompt, alert and dialog
 
Tested on the following devices:
- Dev Alpha B with OS 10.2.1.2141
- Dev Alpha C with OS 10.2.0.1157